Remove discard operator
All terminals are discarded by default, which means nothing is passed
up unless inserted by a raw expression or a rule with an action.

This makes the grammars cleaner and doesn't really change performance
much (although it's now easier for the regex optimizer to build bigger
regexes). If discarding built-up things is necessary in the future, it
can be done with an action.