Improve memoization
Rather than only memoizing at a Rule, now it only happens on a Choice
as that is the only operator that really backtracks (where memoization
makes a difference). Failures, however, are also memoized on Terminals
for the purpose of error reporting.

This commit also adjusts some imports and other meaningful things
found by flake8.