Move mostly to static build config; improve sdist
The sdist now includes the *.c, *.pyx, and *.pxd files. Also, Cython
was removed as a build requirement, which means the sdist can be
installed with the pure-Python sources and won't try to build the
extension modules. In the future it might make sense to build the
extension by default.