Escape special characters in the debug output
When debugging multiline strings I found it a bit difficult to read because the newlines were being printed. Wrapping the snippet in repr fixes this. I then had to add [1:-2] to cut off the quotes that repr introduces.

The only problem this introduces is that if the string contains many special characters it will be more than 12 characters long and hence break the nice formatting.

So I added an additional [:12] on the basis that even though it will sometimes break up the special characters you'll be able to tell that from the line above.